refactor: refactor directory composition
